refactoring: bring package file gathering logic into dedicated module instead of Package.hs : PackageFile, ComponentFile
